Moved countrySelection and toggleNavigationControl below header
