+++ webapp/app/components/countrySelection/components/countryList.less
@@ -6,8 +6,8 @@
 .country-selection-list {
   display: block;
   position: absolute;
-  top: 60px;
-  left: 16px;
+  top: 52px;
+  left: 83px;
   background: white;
   border-radius: 2px;
   box-shadow: @shadow-3;
