+++ webapp/app/components/countrySelection/components/countrySelection.less
@@ -1,17 +1,29 @@
 @import (reference) 'webapp/main/styles/config';
 
 .country-selection {
+  position: relative;
   grid-area: country-selection;
+  display: flex;
+  align-items: center;
+  background-color: @ui-bg;
+  border-bottom: @ui-thin-border;
+  padding: 0 16px;
+}
+
+.btn-toggle-navigation {
+  font-size: 1rem;
+  font-weight: 400;
+}
+
+.btn-country-selection {
   display: flex;
   flex-flow: row nowrap;
   align-items: center;
-  padding: 4px 12px 4px 16px;
-  min-height: @headerHeight;
-  background-color: unset;
-  border-top: unset;
-  border-right: unset;
-  border-left: unset;
-  border-bottom: 1px solid @ui-border;
+  min-width: 320px;
+  padding: 7px 14px;
+  margin-left: 16px;
+  background-color: white;
+  border: @ui-thin-border;
   text-align: left;
   font-size: @font-m;
   transition: @ease-out;
@@ -29,9 +41,8 @@
 .country-selection__flag {
   background-repeat: no-repeat;
   background-size: cover;
-  width: 32px;
-  min-width: 32px;
-  height: 32px;
+  width: 23px;
+  height: 23px;
   box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.15);
   border-radius: 2px;
 }
@@ -47,7 +58,6 @@
   display: block;
   font-weight: 700;
   line-height: 16px;
-  margin-bottom: 4px;
 }
 
 .country-selection__user-role {
