+++ webapp/app/components/header/header.less
@@ -10,22 +10,6 @@
   padding: 0 20px;
 }
 
-.app-header__toggle-navigation-visible {
-  .mixin-header-menu-item;
-  font-size: 1rem;
-  font-weight: 400;
-
-  .label {
-    margin-left: 8px;
-    opacity: 0;
-    transition: opacity 0.1s ease-in;
-
-    &.show {
-      opacity: 1;
-    }
-  }
-}
-
 .app-header__autosave {
   display: inline-block;
   line-height: 16px;
