+++ webapp/app/components/navigation/navigation.less
@@ -9,6 +9,7 @@
 
 .nav {
   &:extend(.nav-base);
+  grid-area: navigation;
   overflow-x: hidden;
   overflow-y: auto;
   height: 100%;
